refactor: removed _repository suffix in mongoldb repository
