+++ src/db/mongoldb/repositories.rs
@@ -1,5 +1,5 @@
-mod user_repository;
-mod chat_repository;
-mod message_repository;
-mod refresh_token_repository;
-mod relation_repository;
+mod user;
+mod chat;
+mod message;
+mod refresh_token;
+mod relation;
